% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{branche}
\alias{branche}
\title{Branche}
\usage{
branche(tree, t)
}
\arguments{
\item{tree}{: Frechet tree}

\item{t}{[numeric]: a node of tree}
}
\description{
Used to prune a maximal Frechet tree
}
\keyword{internal}
