% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{curve.reduc.times}
\alias{curve.reduc.times}
\title{Title}
\usage{
curve.reduc.times(time.init, traj.init, time.new)
}
\arguments{
\item{time.new}{}
}
\description{
Title
}
\keyword{internal}
