% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{DataGenCurves}
\alias{DataGenCurves}
\title{Curves data generation function}
\usage{
DataGenCurves(n)
}
\arguments{
\item{n}{[numeric]: Number of trajectories}
}
\value{
A list with the followig elements: \itemize{
\item \code{X:}
\item \code{Y:}
\item \code{id:}
\item \code{time: }
}
}
\description{
This function allows to ...
}
\examples{
data <- DataGenCurves(50)
}
