% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{elagage}
\alias{elagage}
\title{Pruning Frechet tree}
\usage{
elagage(tree)
}
\arguments{
\item{tree}{: Frechet tree}
}
\description{
To prune a Frechet tree for curves
}
\keyword{internal}
