% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{elagageHuberts}
\alias{elagageHuberts}
\title{Pruning function for Hubert Frechet tree}
\usage{
elagageHuberts(tree)
}
\arguments{
\item{tree}{: Frechet tree}
}
\description{
To prune a tree and selects the optimal sub Frechet tree according to the Huberts statistic
}
\keyword{internal}
