% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{ERimpurity_split}
\alias{ERimpurity_split}
\title{Title}
\usage{
ERimpurity_split(traj, time, id, split, timeScale = 0.1)
}
\arguments{
\item{timeScale}{}
}
\description{
Title
}
\keyword{internal}
