% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechTree.R
\name{FrechetTree}
\alias{FrechetTree}
\title{General Frechet Tree}
\usage{
FrechetTree(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  timeScale = 0.1,
  ncores = NULL,
  FrechetSumOrMax = "max",
  ...
)
}
\arguments{
\item{Curve}{[list]: A list that contains the different input curves. It must
contain the following elements (no choice): \code{X} the matrix of the
different curves, each column code for a different curve variable;
\code{id} is the vector of the identifiers for the different trajectories
contained in \code{X}; \code{time} is the vector of the measurement times
associated with the trajectories contained in \code{X}.}

\item{Scalar}{[list]: A list that contains the different input scalars. It
must contain the following elements (no choice):  \code{X} the matrix of
the scalars, each column code for a different variable; \code{id} is the
vector of the identifiers for each individual.}

\item{Factor}{[list]: A list that contains the different input factors. It
must contain the following elements (no choice):  \code{X} the matrix of
the factors, each column code for a different variable; \code{id} is the
vector of the identifiers for each individual.}

\item{Y}{[list]: A list that contains the output, It must contain the
following elements (no choice): \code{type} defines the nature of the
output, can be "\code{curve}", "\code{sclalar}", "\code{factor}",
"\code{shape}", "\code{image}"; \code{Y} is the output variable; \code{id}
is the vector of the identifiers for each individuals, they should be the
same as the identifiers of the inputs.}

\item{timeScale}{[numeric]: Allow to modify the time scale, increasing or
decreasing the cost of the horizontal shift. If timeScale is very big, then
the Frechet mean tends to the Euclidean distance. If timeScale is very
small, then it tends to the Dynamic Time Warping. Only used when there are
trajectories either in input or output.}

\item{ncores}{[numeric]: Number of cores used to build Frechet randomized
trees in parallel, defaulting to number of cores of the computer minus 1.}

\item{FrechetSumOrMax}{Frechet Mean and Frechet Distance can be define using
the 'sum' function or the 'max' function.}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
General Frechet Tree
}
