% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{Huberts}
\alias{Huberts}
\title{Hubert's statistic for Frechet tree}
\usage{
Huberts(tree)
}
\arguments{
\item{tree}{: Frechet tree}
}
\description{
Computes the Huberts statistic for a given Frechet tree
}
\keyword{internal}
