% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importance_Image.R
\name{Importance_Image}
\alias{Importance_Image}
\title{Optimized Importance variable calculation for servers}
\usage{
Importance_Image(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  range = NULL,
  ncores = NULL,
  timeScale = 0.1
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{Y}{[list]:}

\item{range}{[vector]:}

\item{ncores}{[numeric]:}

\item{timeScale}{[numeric]:}
}
\description{
Optimized Importance variable calculation for servers
}
