% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imp_server.R
\name{Importance_server}
\alias{Importance_server}
\title{Optimized Importance variable calculation for servers}
\usage{
Importance_server(
  Trees,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  type = NULL,
  range = NULL,
  ncores = NULL,
  timeScale = 0.1,
  xerror
)
}
\arguments{
\item{Trees}{[string]:}

\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{Y}{[list]:}

\item{type}{[string]:}

\item{range}{[vector]:}

\item{ncores}{[numeric]:}

\item{timeScale}{[numeric]:}

\item{xerror}{[vector]}
}
\description{
Optimized Importance variable calculation for servers
}
