% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{impurity}
\alias{impurity}
\title{Impurity}
\usage{
impurity(traj, time, id, timeScale = 0.1)
}
\arguments{
\item{traj}{[vector]: Vector of curves}

\item{time}{[vector]: Time measurements of the different curves (same length as traj)}

\item{id}{[vector]: IDs of the different measurements (same length as traj)}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}
}
\description{
Computes the Frechet Variance from a group of curves
}
\keyword{internal}
