% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{impurity_split}
\alias{impurity_split}
\title{Split Impurity}
\usage{
impurity_split(traj, time, id, split, timeScale = 0.1)
}
\arguments{
\item{traj}{[vector]: Trajectories}

\item{time}{[vector]: Times at which measures are made}

\item{id}{[vector]: IDs of the different measurements}

\item{split}{[numeric]: Variable of split}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}
}
\description{
Computes the Frechet variance reduction from a split
}
\keyword{internal}
