% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{noeuds_deg}
\alias{noeuds_deg}
\title{Noeud deg}
\usage{
noeuds_deg(tree)
}
\arguments{
\item{tree}{: Frechet tree}
}
\description{
This function is used to detect the nodes to prune at each step
}
\keyword{internal}
