% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{OOB.rfshape}
\alias{OOB.rfshape}
\title{OOB error for a Frechet random forest}
\usage{
OOB.rfshape(rf, X, Y, time, id, timeScale = 0.1)
}
\arguments{
\item{rf}{: Frechet random forest}

\item{X}{[matrix]: matrix of the explanatory variables used to build the tree}

\item{Y}{[vector]: output curves used to build the tree}

\item{time}{[vector]: time measurements}

\item{id}{[vector]: IDs of measurements}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}
}
\description{
Computes the OOB error of a given Frechet random forest
}
\keyword{internal}
