% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOB_server.R
\name{OOB.server}
\alias{OOB.server}
\title{OOB for random Forest}
\usage{
OOB.server(
  rf,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  ncores = NULL,
  range = NULL,
  Y,
  timeScale = 0.1,
  d_out = 0.1
)
}
\arguments{
\item{rf}{[list]:}

\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{ncores}{[numeric]:}

\item{range}{[vector]:}

\item{Y}{[list]:}

\item{timeScale}{[numeric]}

\item{d_out}{[numeric]}
}
\description{
OOB for random Forest
}
