% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{OOB.tree}
\alias{OOB.tree}
\title{OOB.tree}
\usage{
OOB.tree(tree, X, Y, id, time, timeScale = 0.1)
}
\arguments{
\item{tree}{: Frechet tree}

\item{X}{[matrix]: matrix of the explanatory variables used to build the tree}

\item{Y}{[vector]: output curves used to build the tree}

\item{id}{[vector]: IDs of measurements}

\item{time}{[vector]: time measurements}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}
}
\description{
Computes the oob error of a given tree:
}
\keyword{internal}
