% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{ordonne}
\alias{ordonne}
\title{Ordonne time measurements}
\usage{
ordonne(X, time, id)
}
\arguments{
\item{X}{[vector]: matrix of covariables, each variable must be in column}

\item{time}{[vector]: time measurements of the observations}

\item{id}{[vector]: IDs of the measurements to their belonging curves}
}
\description{
Used to put the data to the right form at a given time
}
\keyword{internal}
