% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\docType{data}
\name{Part.facts}
\alias{Part.facts}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{list} of length 10.
}
\usage{
data(Part.facts)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
\keyword{internal}
