% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{permutation}
\alias{permutation}
\title{Permutation}
\usage{
permutation(Courbes, id)
}
\arguments{
\item{Courbes}{[vector]: Curves, ordered by time measurements}

\item{id}{: IDs of measurements}
}
\description{
This function is used to permute curves in order to compute the OOB error from a Frechet random forest
}
\keyword{internal}
