% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutations.R
\name{permutation_courbes}
\alias{permutation_courbes}
\title{Title}
\usage{
permutation_courbes(Courbes, id)
}
\arguments{
\item{Courbes}{Input Curves to be permuted}

\item{id}{Identifiers of individuals}
}
\description{
Title
}
\keyword{internal}
