% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutations.R
\name{permutation_shapes}
\alias{permutation_shapes}
\title{Title}
\usage{
permutation_shapes(Shapes, id)
}
\arguments{
\item{Shapes}{Input Shapes to be permuted}

\item{id}{Identifiers of individuals}
}
\description{
Title
}
\keyword{internal}
