% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE.R
\name{predRE}
\alias{predRE}
\title{Function to compute individual random effects using hlme output parameters}
\usage{
predRE(model, data)
}
\arguments{
\item{model}{output object from hlme function}

\item{data}{data to compute random effect}
}
\description{
Function to compute individual random effects using hlme output parameters
}
