% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{rf_shape}
\alias{rf_shape}
\title{Frechet random forest}
\usage{
rf_shape(X, Y, id, time, mtry, ntree, timeScale = 0.1, ERtrees = FALSE,
  ...)
}
\arguments{
\item{X}{[matrix]: Matrix of explanatory variables, each column codes for a variable}

\item{Y}{[vector]: Output curves}

\item{id}{[vector]: IDs of measurements}

\item{time}{[vector]: time of measurements}

\item{mtry}{[numeric]: number of variables randomly chosen at each split}

\item{ntree}{[numeric]: number of randomized Frechet trees composing the Frechet random forest}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}

\item{...}{:}
}
\description{
Build a Frechet random Forest
}
\keyword{internal}
