% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{Rtmax}
\alias{Rtmax}
\title{Maximal randomized Fréchet tree}
\usage{
Rtmax(X, Y, id, time, mtry, timeScale = 0.1, ...)
}
\arguments{
\item{X}{[matrix]: Matrix of explanatory variables, each column codes for a variable}

\item{Y}{[vector]: Output curves}

\item{id}{[vector]: IDs of measurements}

\item{time}{[vector]: time measurements}

\item{mtry}{[numeric]: number of variables randomly chosen at each split}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the Euclidean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}

\item{...}{:}
}
\description{
Build a maximal randomized Fréchet tree
}
\keyword{internal}
