% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{Tmax}
\alias{Tmax}
\title{Maximal Frechet tree}
\usage{
Tmax(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  timeScale = 0.1,
  ...
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{Y}{[list]:}

\item{timeScale}{[list]:}

\item{...}{: Optional parameters}
}
\description{
Maximal Frechet tree
}
\keyword{internal}
