% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{TreeShapeCV}
\alias{TreeShapeCV}
\title{Cross validated Frechet tree}
\usage{
TreeShapeCV(X, Y, id, time, ...)
}
\arguments{
\item{X}{[matrix]: a data frame or a matrix of trajectories predictors. Each colunm codes for a trajectory predictor.}

\item{Y}{[vector]: a vector containing the output trajectories (same length as \code{nrow(X)}).}

\item{id}{[vector (factor)]: identifier, one for each trajectory to attribute each measurement of \code{X} and \code{Y} to one of the trajectories (same length as \code{Y}).}

\item{time}{[vector]: time measurements for the observations of both \code{X} and \code{Y} (same length as \code{Y}).}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
Builds a Frechet tree that is selected with cross validation
}
\keyword{internal}
