% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetpc.R
\name{var_split}
\alias{var_split}
\title{Variable split}
\usage{
var_split(X, Y, time, id, timeScale = 0.1, ...)
}
\arguments{
\item{X}{[matrix]: Matrix of explanatory variables, each column codes for a variable.}

\item{Y}{[vector]: Output curves}

\item{time}{[vector]: Time at which measures are made}

\item{id}{[vector]: IDs of trajectories}

\item{timeScale}{[numeric]: allow to modify the time scale, increasing or decreasing the cost of the horizontal shift. If timeScale is very big, then the Frechet mean tends to the euclidiean distance. If timeScale is very small, then it tends to the Dynamic Time Warping.}

\item{...}{:}
}
\description{
Determines the split from a set of input and output variables which are curves:
}
\keyword{internal}
