% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{var_split_summary}
\alias{var_split_summary}
\title{Antho Variable Split function}
\usage{
var_split_summary(X, Y, timeScale = 0.1, nsplit_option = NULL,
  nodesize = 5)
}
\arguments{
\item{timeScale}{}
}
\description{
Antho Variable Split function
}
\keyword{internal}
