% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves_reduct.R
\name{Curve.reduc.times}
\alias{Curve.reduc.times}
\title{Title}
\usage{
Curve.reduc.times(time.init, traj.init, time.new)
}
\arguments{
\item{time.init}{Initial time values of the trajectory}

\item{traj.init}{Initial trajectory values}

\item{time.new}{New times values for the output trajectory}
}
\description{
Title
}
\keyword{internal}
