% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{ERvar_split}
\alias{ERvar_split}
\title{Extremely randomized split}
\usage{
ERvar_split(X, Y, ntry = 3, fast = FALSE, timeScale = 0.1)
}
\arguments{
\item{ntry}{}
}
\description{
Extremely randomized split
}
\keyword{internal}
