% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{FrechetTree}
\alias{FrechetTree}
\title{General Frechet Tree}
\usage{
FrechetTree(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Y,
  timeScale = 0.1,
  ncores = NULL
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Y}{[list]:}

\item{timeScale}{[numeric]:}

\item{ncores}{[numeric]:}
}
\value{

}
\description{
General Frechet Tree
}
