% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{impurity}
\alias{impurity}
\title{Impurity}
\usage{
impurity(Y, center = NULL, timeScale = 0.1)
}
\arguments{
\item{center}{}
}
\description{
Compute the impurity of a given vector
}
\keyword{internal}
