% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impurity.R
\name{impurity}
\alias{impurity}
\title{Impurity}
\usage{
impurity(Y, d_out = 0.1, FrechetSumOrMax = "max", ...)
}
\arguments{
\item{Y}{[list]: A list that contains the output, It must contain the
following elements (no choice): \code{type} defines the nature of the
output, can be "\code{curve}", "\code{sclalar}", "\code{factor}",
"\code{shape}", "\code{image}"; \code{Y} is the output variable; \code{id}
is the vector of the identifiers for each individuals, they should be the
same as the identifiers of the inputs.}

\item{d_out}{[numeric]: Time scale for the output curves (\code{d_out=0.1} by
default).}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
Compute the impurity of a given vector
}
\keyword{internal}
