% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impurity_split.R
\name{impurity_split}
\alias{impurity_split}
\title{Impurity Split}
\usage{
impurity_split(Y, split, timeScale = 0.1, ...)
}
\arguments{
\item{Y}{[list]: A list that contains the output, It must contain the
following elements (no choice): \code{type} defines the nature of the
output, can be "\code{curve}", "\code{sclalar}", "\code{factor}",
"\code{shape}", "\code{image}"; \code{Y} is the output variable; \code{id}
is the vector of the identifiers for each individuals, they should be the
same as the identifiers of the inputs.}

\item{split}{}

\item{timeScale}{[numeric]: Allow to modify the time scale, increasing or
decreasing the cost of the horizontal shift. If timeScale is very big, then
the Frechet mean tends to the Euclidean distance. If timeScale is very
small, then it tends to the Dynamic Time Warping. Only used when there are
trajectories either in input or output.}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
Impurity Split
}
\keyword{internal}
