% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{OOB.rfshape}
\alias{OOB.rfshape}
\title{OOB for random Forest}
\usage{
OOB.rfshape(rf, Curve = NULL, Scalar = NULL, Factor = NULL,
  Shape = NULL, Image = NULL, Y, timeScale = 0.1)
}
\arguments{
\item{timeScale}{}
}
\description{
OOB for random Forest
}
\keyword{internal}
