% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOB_tree.R
\name{OOB.tree}
\alias{OOB.tree}
\title{OOB tree}
\usage{
OOB.tree(
  tree,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  timeScale = 0.1,
  d_out = 0.1
)
}
\arguments{
\item{tree}{[list]: Frechet tree.}

\item{Curve}{[list]: A list that contains the input curves.}

\item{Scalar}{[list]: A list that contains the input scalars.}

\item{Factor}{[list]: A list that contains the input factors.}

\item{Shape}{[list]: A list that contains the input shape.}

\item{Image}{[list]: A list that contains the input images.}

\item{Y}{[list]: Output.}

\item{timeScale}{[numeric]: Time scale for the input  curves (\code{timeScale=0.1} by default)}

\item{d_out}{[numeric]: Time scale for the input curves (\code{d_out=0.1} by default)}
}
\description{
OOB tree
}
