% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{OOB.tree}
\alias{OOB.tree}
\title{OOB tree}
\usage{
OOB.tree(
  tree,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  timeScale = 0.1
)
}
\arguments{
\item{timeScale}{}
}
\description{
OOB tree
}
\keyword{internal}
