% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{predict.FrechForest}
\alias{predict.FrechForest}
\title{Predict with Frechet random forests}
\usage{
\method{predict}{FrechForest}(object, Curve = NULL, Scalar = NULL,
  Factor = NULL, Shape = NULL, Image = NULL, aligned.shape = FALSE,
  timeScale = 0.1, ...)
}
\arguments{
\item{object}{: Frechet random forest}

\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Shape}{[list]:}

\item{Image}{[list]:}

\item{aligned.shape}{[logical]:}

\item{timeScale}{[numeric]:}

\item{...}{: optional parameters to be passed to the low level function}
}
\description{
Predict with Frechet random forests
}
