% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{rf_shape_para}
\alias{rf_shape_para}
\title{Parallelized Frechet random Forest}
\usage{
rf_shape_para(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  mtry,
  ntree,
  ncores,
  ERT = FALSE,
  ntry = 3,
  nodesize = 5,
  timeScale = 0.1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Parallelized Frechet random Forest
}
\keyword{internal}
