% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrechForest.R
\name{var_split}
\alias{var_split}
\title{Classical Variable Split function}
\usage{
var_split(X, Y, timeScale = 0.1)
}
\arguments{
\item{timeScale}{}
}
\description{
Classical Variable Split function
}
\keyword{internal}
